# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CROAS OS — Sistema de Gestão 360 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [commercial]
      summary: List leads with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
            enum: [lead, call_agendada, call_realizada, proposta, fechado, perdido]
        - name: origin
          in: query
          schema:
            type: string
            enum: [instagram, indicacao, parceria, palestra, outro]
        - name: assignedToId
          in: query
          schema:
            type: number
        - name: followUpToday
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [commercial]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [commercial]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [commercial]
      summary: Update a lead (name, notes, next follow-up, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [commercial]
      summary: Delete a lead (MASTER only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /leads/{id}/move-stage:
    post:
      operationId: moveLeadStage
      tags: [commercial]
      summary: Move a lead to a new pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveStageInput"
      responses:
        "200":
          description: Updated lead with new stage (if stage=fechado, convertedStudentId is populated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Invalid stage transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── COMMERCIAL — GOALS ───────────────────────────────────────────────────────
  /commercial/goals/{year}/{month}:
    get:
      operationId: getCommercialGoal
      tags: [commercial]
      summary: Get the monthly commercial goal
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: number
        - name: month
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Monthly goal (null if not set)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommercialGoal"
    put:
      operationId: setCommercialGoal
      tags: [commercial]
      summary: Create or update the monthly commercial goal
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: number
        - name: month
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommercialGoalInput"
      responses:
        "200":
          description: Upserted goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommercialGoal"

  # ─── COMMERCIAL — ANALYTICS ───────────────────────────────────────────────────
  /commercial/analytics/funnel:
    get:
      operationId: getConversionFunnel
      tags: [commercial]
      summary: Conversion funnel — count and rate per stage
      parameters:
        - name: year
          in: query
          schema:
            type: number
        - name: month
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Funnel data per stage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FunnelStageData"
  /commercial/analytics/by-origin:
    get:
      operationId: getConversionByOrigin
      tags: [commercial]
      summary: Lead count and conversion rate per origin channel
      parameters:
        - name: year
          in: query
          schema:
            type: number
        - name: month
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Data per origin
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OriginData"

  # ─── FINANCE — CATEGORIES ────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Lead:
      type: object
      required: [id, fullName, origin, stage, isActive, createdAt, updatedAt]
      properties:
        id:
          type: number
        fullName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        origin:
          type: string
          enum: [instagram, indicacao, parceria, palestra, outro]
        stage:
          type: string
          enum: [lead, call_agendada, call_realizada, proposta, fechado, perdido]
        assignedToId:
          type: ["number", "null"]
        assignedToName:
          type: ["string", "null"]
        lastContactAt:
          type: ["string", "null"]
          format: date-time
        nextFollowUpAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        proposalValue:
          type: ["number", "null"]
        isActive:
          type: boolean
        convertedStudentId:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LeadInput:
      type: object
      required: [fullName]
      properties:
        fullName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        origin:
          type: string
          enum: [instagram, indicacao, parceria, palestra, outro]
        stage:
          type: string
          enum: [lead, call_agendada, call_realizada, proposta, fechado, perdido]
        assignedToId:
          type: ["number", "null"]
        nextFollowUpAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        proposalValue:
          type: ["number", "null"]
    LeadUpdate:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        origin:
          type: string
          enum: [instagram, indicacao, parceria, palestra, outro]
        assignedToId:
          type: ["number", "null"]
        lastContactAt:
          type: ["string", "null"]
        nextFollowUpAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        proposalValue:
          type: ["number", "null"]
        isActive:
          type: boolean
    MoveStageInput:
      type: object
      required: [stage]
      properties:
        stage:
          type: string
          enum: [lead, call_agendada, call_realizada, proposta, fechado, perdido]
        notes:
          type: ["string", "null"]
        # Required when moving to "fechado" — plan to enroll student in
        planId:
          type: ["number", "null"]
    CommercialGoal:
      type: object
      required: [id, year, month, targetDeals, targetRevenue, createdAt, updatedAt]
      properties:
        id:
          type: number
        year:
          type: number
        month:
          type: number
        targetDeals:
          type: number
        targetRevenue:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        # Computed actuals
        closedDeals:
          type: number
        closedRevenue:
          type: number
    CommercialGoalInput:
      type: object
      required: [targetDeals, targetRevenue]
      properties:
        targetDeals:
          type: number
        targetRevenue:
          type: number
        notes:
          type: ["string", "null"]
    FunnelStageData:
      type: object
      required: [stage, count, conversionRate]
      properties:
        stage:
          type: string
          enum: [lead, call_agendada, call_realizada, proposta, fechado, perdido]
        count:
          type: number
        conversionRate:
          type: number
    OriginData:
      type: object
      required: [origin, total, closed, conversionRate]
      properties:
        origin:
          type: string
          enum: [instagram, indicacao, parceria, palestra, outro]
        total:
          type: number
        closed:
          type: number
        conversionRate:
          type: number

    # ─── FINANCE ──────────────────────────────────────────────────────────────
...